# docs/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "This is the admin API for the Urchin app.",
        "title": "GoCMS Admin API",
        "contact": {
            "name": "Ricardo",
            "email": "ricardobenthem@gmail.com"
        },
        "license": {
            "name": "MIT"
        },
        "version": "1.0.0"
    },
    "host": "localhost:8081",
    "basePath": "/",
    "paths": {
...
        "/card_schema": {
            "get": {
                "description": "Retrieves a paginated list of card schemas.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "card_schema"
                ],
                "summary": "Get list of card schemas",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 0,
                        "description": "Pagination offset",
                        "name": "offset",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 0,
                        "description": "Pagination limit (0 means no limit)",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/admin_app.GetSchemaasResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid query parameters",
                        "schema": {
                            "$ref": "#/definitions/common.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Server error",
                        "schema": {
                            "$ref": "#/definitions/common.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "description": "Creates a new card schema by validating and storing the provided JSON schema and title.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "card_schema"
                ],
                "summary": "Add a new card schema",
                "parameters": [
                    {
                        "description": "Card schema to add",
                        "name": "schema",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/admin_app.AddCardSchemaRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/admin_app.CardIdResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid request body or schema",
                        "schema": {
                            "$ref": "#/definitions/common.ErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "description": "Deletes a card schema by its ID.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "card_schema"
                ],
                "summary": "Delete a card schema",
                "parameters": [
                    {
                        "description": "ID of the card schema to delete",
                        "name": "schema",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/admin_app.DeleteSchemaBinding"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Deleted schema ID",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid request or deletion error",
                        "schema": {
                            "$ref": "#/definitions/common.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/card_schema/{id}": {
            "get": {
                "description": "Retrieves a card schema by its UUID.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "card_schema"
                ],
                "summary": "Get a card schema by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Card schema UUID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/common.CardSchema"
                        }
                    },
                    "400": {
                        "description": "Invalid schema ID",
                        "schema": {
                            "$ref": "#/definitions/common.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Schema not found",
                        "schema": {
                            "$ref": "#/definitions/common.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "admin_app.AddCardSchemaRequest": {
            "type": "object",
            "properties": {
                "schema": {
                    "description": "JSON schema string\nin: body\nrequired: true",
                    "type": "string"
                },
                "title": {
                    "description": "Title of the card schema\nin: body\nrequired: true",
                    "type": "string"
                }
            }
        },
...
        "admin_app.CardIdResponse": {
            "type": "object",
            "properties": {
                "id": {
                    "description": "ID of the card",
                    "type": "string"
                }
            }
        },
...
        "admin_app.DeleteSchemaBinding": {
            "type": "object",
            "required": [
                "id"
            ],
            "properties": {
                "id": {
                    "description": "UUID of the schema to delete\nin: path\nrequired: true",
                    "type": "string"
                }
            }
        },
...
        "admin_app.GetSchemaasResponse": {
            "type": "object",
            "properties": {
                "schemas": {
                    "description": "List of card schemas",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/common.CardSchema"
                    }
                }
            }
        },
...
        "common.CardSchema": {
            "type": "object",
            "properties": {
                "cards": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "schema": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                },
                "uuid": {
                    "type": "string"
                }
            }
        },
        "common.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                },
                "msg": {
                    "type": "string"
                }
            }
        },
...
    }
}